Point bridge to experimental fix to sync ntt error
The bridge was running into this issue every few days and it is likely
due to a networking or internal error of some sort inside or between the
native haskell clients. This new cardano-deps branch attempts to fix
this by restarting the syncing up to a threshold when such error occurs
and with more logging info. If this does not resolve the issue, a
higher-level restart within the bridge of all network refreshers might
be required.